feat (refs T31120): Move updateRecommendations to SegmentBulkEditorService
